requirement.txt
django
opencv-python
ultralytics
djangorestframework
django-cors-headers